fix: resolve 6 known interpreter bugs (sort named-type, fmt.Stringer, %T, %v _gig_id, strconv narrowing)
Bug 1: Add OpChangeType opcode for named-type conversion (e.g. []int → sort.IntSlice)
  - bytecode/opcode.go: new OpChangeType with [type_idx:2][src_local:2] operands
  - compiler/compile_value.go: emit OpChangeType for SSA ChangeType instructions
  - vm/ops_dispatch.go: OpChangeType handler with source-local aliasing

Bug 1b: sort.Ints in-place writeback for []int64 internal representation
  - stdlib/packages/sort.go: direct_sort_Ints with []int64→[]int writeback

Bug 3: fmt.Stringer interface honored on interpreted structs
  - value/accessor.go: MethodResolver cross-package callback for compiled methods
  - vm/vm.go: resolveCompiledMethod() executes compiled String() via temp VM
  - stdlib/packages/fmt.go: gigStructFormatter calls String() if available

Bug 4: fmt.Sprintf(%T) reports correct type name
  - vm/typeconv.go: encode qualified name in _gig_id PkgPath (#PkgName.TypeName)
  - vm/call.go: update inferReceiverTypeName for qualified PkgPath format
  - stdlib/packages/fmt.go: sprintfWithTypeAwareness() intercepts %T verbs

Bug 5: Hide _gig_id sentinel field from %v output
  - stdlib/packages/fmt.go: gigStructFormatter.Format() skips _gig_id

Bug 6: strconv.ParseInt/ParseUint return int64/uint64 (not narrowed int/uint)
  - stdlib/packages/strconv.go: MakeInt64/MakeUint64 in DirectCall wrappers

All 10 known-issue sub-tests pass. Full regression suite green.

+++ tests/thirdparty_interp_test.go
@@ -110,8 +110,8 @@ func TestCorrectnessThirdparty(t *testing.T) {
 	runCategory(t, "Strconv", srcStrconv, []thirdpartyCase{
 		{"StrconvParseBool", 1},
 		{"StrconvFormatBool", "false"},
-		{"StrconvParseInt", 12345},
-		{"StrconvParseUint", uint(12345)},
+		{"StrconvParseInt", int64(12345)},
+		{"StrconvParseUint", uint64(12345)},
 		{"StrconvFormatInt", "-12345"},
 		{"StrconvFormatUint", "12345"},
 		{"StrconvParseFloat", float64(123.45)},
